+++ tests/test_end_to_end/golden-master/my_test_api_client/models/abc_response.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from datetime import datetime
 from typing import Any, Dict, List, Optional, cast